save methods pushed up to AbstractGraphRepository. Neo4jTemplate.postConstruct declared as initMethod. Annotated the repository implementations with @Repository. Unnecessary component-scan removed
