+++ spring-data-neo4j-cross-store/src/test/resources/org/springframework/data/neo4j/partial/Neo4jGraphRecommendationTest-context.xml
@@ -38,34 +38,6 @@
 	-->
 	<context:spring-configured/>
 
-	<!--
-		This declaration will cause Spring to locate every @Component,
-		@Repository and @Service in your application. In practical terms this
-		allows you to write a POJO and then simply annotate the new POJO as an
-		@Service and Spring will automatically detect, instantiate and
-		dependency inject your service at startup time. Importantly, you can
-		then also have your new service injected into any other class that
-		requires it simply by declaring a field for your service inside the
-		relying class and Spring will inject it. Note that two exclude filters
-		are declared. The first ensures that Spring doesn't spend time
-		introspecting Roo-specific ITD aspects. The second ensures Roo doesn't
-		instantiate your @Controller classes, as these should be instantiated
-		by a web tier application context. Refer to web.xml for more details
-		about the web tier application context setup services.
-		
-		Furthermore, this turns on @Autowired, @PostConstruct etc support. These 
-		annotations allow you to use common Spring and Java Enterprise Edition 
-		annotations in your classes without needing to do any special configuration. 
-		The most commonly used annotation is @Autowired, which instructs Spring to
-		dependency inject an object into your class.
-	-->
-	<context:component-scan base-package="org.springframework.data.neo4j">
-		<context:exclude-filter expression=".*_Roo_.*" type="regex"/>
-		<context:exclude-filter expression="org.springframework.stereotype.Controller" type="annotation"/>
-		<context:exclude-filter expression="org.springframework.context.annotation.Configuration" type="annotation"/>
-	</context:component-scan>
-	
-	
 	<!-- 
 	    Start datastore config
     -->
